services/api-gateway/requirements.txt
fastapi==0.104.1
httpx==0.25.2
PyJWT==2.8.0
slowapi==0.1.8
python-dotenv==1.0.0
prometheus-client==0.19.0
structlog==23.2.0 